requirements.txt
python-dotenv==1.0.0
pymongo==4.6.1
razorpay==1.4.0
cryptography==41.0.7 